started updating directory structure for managing access switched to yarn from npm general re-examination and refactoring
