--- .nvm
@@ -1,1 +1,0 @@
-8.6.0
