+++ .nvmrc
@@ -1,1 +1,1 @@
-8.6.0
+10.15.0
