+++ plugin/src/ramdaExt/index.js
@@ -1,0 +1,9 @@
+const R = require('ramda');
+
+const first = R.nth(0);
+const second = R.nth(1);
+
+module.exports = {
+	first,
+	second
+}
