+++ reference/json-ld/package.json
@@ -4,7 +4,9 @@
   "description": "reference schema site using metalsmith-ldschema",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "build:debug": "node inspect ./build.js",
+    "build": "node ./build.js"
   },
   "repository": {
     "type": "git",
